Optimized Order and Category repositories for improved performance

+++ Shop.Application/Category/Commands/DeleteCategory/DeleteCategoryCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using MediatR;
+using Shop.Application.Exceptions;
 using Shop.Domain.Interfaces;
 
 namespace Shop.Application.Category.Commands.DeleteCategory
@@ -15,6 +16,11 @@ namespace Shop.Application.Category.Commands.DeleteCategory
         public async Task<Unit> Handle(DeleteCategoryCommand request, CancellationToken cancellationToken)
         {
             var category = await _categoryRepository.GetByEncodedName(request.EncodedName);
+
+            if (category == null)
+            {
+                throw new NotFoundException("Category not found.");
+            }
             
             await _categoryRepository.Delete(category);
